watchdogs/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-eventgrid
beautifulsoup4
azure-data-tables
azure-keyvault-secrets
azure-identity